* Fixed all signed/unsigned mismatch warnings * Renamed SLADELog::DoLog to avoid incorrect deprecation warnings from wxLog::DoLog.
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@394 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/Console.h
@@ -8,7 +8,7 @@ class ConsoleCommand {
 private:
 	string name;
 	void (*commandFunc)(vector<string>);
-	int min_args;
+	size_t min_args;
 
 public:
 	ConsoleCommand(string name, void(*commandFunc)(vector<string>), int min_args);
@@ -43,7 +43,7 @@ public:
 	}
 
 	int numCommands() { return (int) commands.size(); }
-	ConsoleCommand& command(int index);
+	ConsoleCommand& command(size_t index);
 
 	void addCommand(ConsoleCommand &c);
 	void execute(string command);
